requirements.txt
WMI==1.5.1
py-cpuinfo==9.0.0
termcolor==2.4.0
typer==0.12.3
pyinstaller==6.10.0